+++ test/sql/install.sql
@@ -24,7 +24,7 @@ select relname from pg_catalog.pg_class,
   order by 1;
 
 -- check the emaj_version_hist content
-select verh_version from emaj.emaj_version_hist;
+select verh_version, verh_installed_by_superuser from emaj.emaj_version_hist;
 select emaj.emaj_get_version();
 
 -- check history