+++ test/sql/non_superuser_install.sql
@@ -7,7 +7,7 @@ SET client_min_messages TO WARNING;
 -- Create roles, give grants and create needed extensions
 ------------------------------------------------------------
 -- _regress_emaj_install owns the emaj environment
-create role _regress_emaj_install login password 'install';
+create role _regress_emaj_install login password 'install' createrole;
 
 -- _regress_emaj_app owns the application objects
 create role _regress_emaj_app login password 'app';
@@ -17,6 +17,8 @@ grant create on schema public to _regress_emaj_install;
 
 create extension dblink;
 
+----\du+
+
 ------------------------------------------------------------
 -- Create the application objects and give grants to _regress_emaj_install
 ------------------------------------------------------------
@@ -24,6 +26,7 @@ set role _regress_emaj_app;
 
 DROP SCHEMA IF EXISTS mySchema7 CASCADE;
 CREATE SCHEMA mySchema7;
+GRANT USAGE ON SCHEMA mySchema7 TO _regress_emaj_install;
 
 SET search_path=mySchema7;
 
@@ -35,6 +38,9 @@ CREATE TABLE myTbl1 (
 
 CREATE SEQUENCE mySeq1;
 
+GRANT ALL ON ALL TABLES IN SCHEMA mySchema7 TO _regress_emaj_install;
+GRANT ALL ON ALL SEQUENCES IN SCHEMA mySchema7 TO _regress_emaj_install;
+
 reset search_path;
 
 ------------------------------------------------------------
@@ -43,13 +49,29 @@ reset search_path;
 set role _regress_emaj_install;
 
 \set ECHO errors
-
 \i sql/emaj-devel.sql
-
 \set ECHO all
 
-\dx
+-- check the emaj_version_hist content
+select verh_version, verh_installed_by_superuser from emaj.emaj_version_hist;
+select emaj.emaj_get_version();
 
+------------------------------------------------------------
+-- build a tables group
+------------------------------------------------------------
+select emaj.emaj_create_group('myGroup7');
+----select emaj.emaj_assign_table('myschema7', 'mytbl1', 'myGroup7');
+select emaj.emaj_assign_sequence('myschema7', 'myseq1', 'myGroup7');
+
+------------------------------------------------------------
+-- start the group and perform data changes
+------------------------------------------------------------
+select emaj.emaj_start_group('myGroup7', 'M1');
+
+
+
+select emaj.emaj_disable_protection_by_event_triggers();
+select emaj.emaj_enable_protection_by_event_triggers();
 
 ------------------------------------------------------------
 -- drop the extension