+++ test/sql/upgrade_while_logging.sql
@@ -34,7 +34,7 @@ select relname from pg_catalog.pg_class,
   order by 1;
 
 -- check the emaj_version_hist content
-select verh_version from emaj.emaj_version_hist order by verh_time_range;
+select verh_version, verh_installed_by_superuser from emaj.emaj_version_hist order by verh_time_range;
 select emaj.emaj_get_version();
 
 -- check tables list